Add LootBox to Dependencies.cs
