+++ Assets/Scripts/Meta/LootBox/LootBoxCallerTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 110ba71cc185b7e409c5afc286a6def8
+guid: 8f5447f60a95bfe4289234b56156c40e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
