+++ Assets/Scripts/Utility/Dependencies.cs
@@ -1,7 +1,7 @@
 using Meta.CardSystem;
 using Meta.Interfaces;
+using Meta.LootBox;
 using UnityEngine;
-
 
 namespace Utility
 {
@@ -11,7 +11,7 @@ namespace Utility
         public readonly ISettableCardHand PlayerCardHand = new PlayerCardHand();
         public readonly INormalCoinCarrier NormalCoinCarrier;
         public readonly IPremiumCoinCarrier PremiumCoinCarrier;
-        public readonly ILootBoxSystem LootBoxSystem;
+        public readonly ILootBoxSystem LootBoxSystem = new LootBox();
         
         
         public static Dependencies Instance { get; private set; }
